requirements.txt
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-PyMongo==2.3.0
gevent==23.9.1
greenlet==3.0.0rc3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
load-dotenv==0.1.0
MarkupSafe==2.1.3
pycparser==2.21
PyJWT==2.8.0
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.4
Werkzeug==2.3.7
WSGIserver==1.3
zope.event==5.0
zope.interface==6.0